requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.5
django-baton==5.0.2
django-environ==0.12.0
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
